Rename logprobs module into probs.
