+++ src/stats/probs.rs
